test/site/pip-requirements.txt
autoflake==1.2
autopep8==1.4
Babel==2.6.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.15
django-filter==2.0.0
django-phonenumber-field==2.0.1
djangorestframework==3.8.2
docformatter==1.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.2.1
MarkupSafe==1.0
phonenumbers==8.9.13
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pyflakes==2.0.0
pyformat==0.7
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
six==1.11.0
transliterate==1.10.2
unify==0.4
untokenize==0.1.1
uritemplate==3.0.0
urllib3==1.23
